+++ api/billing/routes.go
@@ -1,0 +1,11 @@
+package billing
+
+import (
+	"github.com/gin-gonic/gin"
+)
+
+func SetupRoutes(r *gin.RouterGroup) {
+	r.POST("/", handlePOSTBillingAccount)
+	r.GET("/:id", handleGETBillingAccount)
+	r.DELETE("/:id", handleDELETEBillingAccount)
+}
