+++ api/users/apps.go
@@ -1,0 +1,26 @@
+package users
+
+import (
+	"strconv"
+
+	"github.com/gin-gonic/gin"
+	"github.com/hackclub/hack-as-a-service/db"
+	"gorm.io/gorm"
+)
+
+func handleGETUserApps(c *gin.Context) {
+	_db := c.MustGet("db").(*gorm.DB)
+	id, err := strconv.Atoi(c.Params.ByName("id"))
+	if err != nil {
+		c.JSON(400, gin.H{"status": "error", "message": "Invalid user ID"})
+		return
+	}
+
+	var apps []db.App
+	result := _db.Raw("SELECT apps.* FROM apps INNER JOIN user_apps ON user_apps.app_id = apps.id WHERE user_apps.user_id = ?", uint(id)).Scan(&apps)
+	if result.Error != nil {
+		c.JSON(500, gin.H{"status": "error", "message": result.Error})
+	} else {
+		c.JSON(200, gin.H{"status": "ok", "apps": apps})
+	}
+}
