+++ api/users/get.go
@@ -1,0 +1,26 @@
+package users
+
+import (
+	"strconv"
+
+	"github.com/gin-gonic/gin"
+	"github.com/hackclub/hack-as-a-service/db"
+	"gorm.io/gorm"
+)
+
+func handleGETUser(c *gin.Context) {
+	_db := c.MustGet("db").(*gorm.DB)
+	id, err := strconv.Atoi(c.Params.ByName("id"))
+	if err != nil {
+		c.JSON(400, gin.H{"status": "error", "message": "Invalid user ID"})
+		return
+	}
+
+	var user db.User
+	result := _db.First(&user, "id = ?", id)
+	if result.Error != nil {
+		c.JSON(500, gin.H{"status": "error", "message": result.Error})
+	} else {
+		c.JSON(200, gin.H{"status": "ok", "user": user})
+	}
+}
