+++ api/users/routes.go
@@ -1,0 +1,12 @@
+package users
+
+import (
+	"github.com/gin-gonic/gin"
+)
+
+func SetupRoutes(r *gin.RouterGroup) {
+	r.POST("/", handlePOSTUser)
+	r.GET("/:id", handleGETUser)
+	r.DELETE("/:id", handleDELETEUser)
+	r.GET("/:id/apps", handleGETUserApps)
+}
