--- db/db.go
@@ -1,7 +1,0 @@
-package db
-
-import (
-	"gorm.io/gorm"
-)
-
-var db *gorm.DB
